Clean up homepage
